# V2 chess engine built in Rust with 100x performance improvement