--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,9 @@
+# Rust binaries
+/target
+
+# Scripts
+__script.py
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]