--- a/gym_chess/__init__.py
+++ b/gym_chess/__init__.py
@@ -1,3 +1,7 @@
+# rust module
+from gym_chess.gym_chess import ChessEngine
+
+# register envs
 from gym.envs.registration import register
 
 register(
@@ -23,5 +27,17 @@ register(
 register(
     id="ChessVsSelf-v1",
     entry_point="gym_chess.envs:ChessEnvV1",
+    kwargs={"opponent": "none"},
+)
+
+register(
+    id="ChessVsRandomBot-v2",
+    entry_point="gym_chess.envs:ChessEnvV2",
+    kwargs={"opponent": "random"},
+)
+
+register(
+    id="ChessVsSelf-v2",
+    entry_point="gym_chess.envs:ChessEnvV2",
     kwargs={"opponent": "none"},
 )