--- a/gym_chess/envs/__init__.py
+++ b/gym_chess/envs/__init__.py
@@ -1,2 +1,3 @@
 from gym_chess.envs.chess_v0 import ChessEnvV0
 from gym_chess.envs.chess_v1 import ChessEnvV1
+from gym_chess.envs.chess_v2 import ChessEnvV2