--- a/gym_chess/envs/chess_v1.py
+++ b/gym_chess/envs/chess_v1.py
@@ -287,6 +287,18 @@ class ChessEnvV1(gym.Env):
         return other_player
 
     @property
+    def possible_moves(self):
+        return self._possible_moves
+
+    @possible_moves.setter
+    def possible_moves(self, moves):
+        self._possible_moves = moves
+
+    @property
+    def possible_actions(self):
+        return [self.move_to_action(m) for m in self.possible_moves]
+
+    @property
     def info(self):
         return dict(
             state=self.state,