--- a/lint
+++ b/lint
@@ -1,2 +1,3 @@
 #! /bin/bash
+black -l 100 .
 cargo fmt