--- a/setup.py
+++ b/setup.py
@@ -21,7 +21,7 @@ install_requires = []
 
 setup(
     name="gym_chess",
-    version="0.2",
+    version="0.3",
     classifiers=[
         "License :: OSI Approved :: MIT License",
         "Development Status :: 3 - Alpha",