# `ChessEnvV2`: a chess environment for gym built in Rust ! 

The `v2` environment uses a chess engine implemented in Rust that uses [PyO3](https://github.com/PyO3/pyo3) to bind to the Python interpreter. Rust is an amazing compiled language and this project holds 2 configurations:

- `Cargo.py.toml` is used to build the library into a Python module with `setup.py`
- `Cargo.dev.toml` is used to build directly with `cargo` and to access the library in the `main.rs` script for development purposes

Note: I haven't found a way to specify the Cargo toml file to either process, so copy the contents of the config you want to use into `Cargo.toml` to make it work.


## Installation

To build as a python module:

```
pip install -e .
```

To run as a Rust library, copy the contents of `Cargo.dev.toml` into `Cargo.toml` and use cargo to run the `main.rs` file.

```
cargo build
cargo run
```


## The chess engine

`ChessEnvV2` imports the Rust chess engine and keeps it around in `env.engine`

```python

from gym_chess import ChessEngine

engine = ChessEngine()

# moves are returned as strings with regular moves having the "e2e4" format
# and castle moves as constant string e.g. "CASTLE_KING_SIDE_WHITE"
moves = engine.get_possible_moves(state: dict, player: str, attack: bool)
castle_moves = engine.get_possible_moves(state: dict, player: str)

# both state and new_state are python dictionaries
new_state = engine.next_state(state: dict, player, move)

# this will update other aspects of the state for a given board position
# e.g. if the player can castle or if the king is checked
state = engine.update_state(state: dict)

```

The state is a dictionary that holds the board positions in addition to other useful pieces of state. It's the only major difference with the `v1` class. The `state` holds the following information:

- `board`: 8x8 array with the piece positions
- `current_player`: "WHITE" or "BLACK"
- `white_king_castle_is_possible`: boolean
-  `white_queen_castle_is_possible`: boolean
- `black_king_castle_is_possible`: boolean
- `black_queen_castle_is_possible`: boolean
- `white_king_is_checked`: boolean
- `black_king_is_checked`: boolean


## Benchmarks

The `v2` environment is over 100 times faster than the `v1` environment. However, since most of the code is written in Rust, it's generally harder to debug.

```python

from gym_chess.envs import ChessEnvV1, ChessEnvV2

env_v1 = ChessEnvV1()
env_v2 = ChessEnvV2()

# v1: written in Python
>>> %timeit -n 50 -r 8 env_v1.get_possible_moves()
## 29.5 ms ± 872 µs per loop (mean ± std. dev. of 8 runs, 50 loops each)

# v2: compiled in Rust
>>> %timeit -n 50 -r 8 env_v2.get_possible_moves()
## 240 µs ± 31.9 µs per loop (mean ± std. dev. of 8 runs, 50 loops each)

```

## Tests

Tests have been added to `gym_chess/tests/v2` and can be run with `pytest`.